fix: change int dtypes to size_t for doc id and word occurrence refactor unordered_map find and bracket operators update map and unordered_map tests change the required number of cmd line args

+++ index.cc
@@ -4,7 +4,7 @@
 
 #include "index.h"
 
-void Index::Add(std::string &word, int id) {
+void Index::Add(std::string &word, size_t id) {
   index_[word][id]++;
 }
